+++ sdl3-sys/ChangeLog.md
@@ -6,6 +6,7 @@
     - Remove deprecated aliases
     - Change return type of `SDL_Vulkan_GetVkGetInstanceProcAddr`
     - Make some functions safe and/or const
+    - Add `SDL_Event::event_type()`
     - MSRV 1.85
 
 - 0.5.4: